+++ src/core/AudioSummingJunction.cpp
@@ -52,8 +52,9 @@ bool AudioSummingJunction::isConnected(std::shared_ptr<AudioNodeOutput> o) const
     return false;
 }
 
-size_t AudioSummingJunction::numberOfRenderingConnections(ContextRenderLock&) const {
-    size_t count = 0;
+uint32_t AudioSummingJunction::numberOfRenderingConnections(ContextRenderLock&) const 
+{
+    uint32_t count = 0;
     for (auto i : m_renderingOutputs) {
         if (!i.expired())
             ++count;
